requirements.txt
# Open-A2A Core Dependencies
nats-py>=2.6.0
